Move symfony bundle into dedicated namespace
