+++ src/Bundle/DependencyInjection/SourceabilityInstrumentationExtension.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sourceability\InstrumentationBundle\DependencyInjection;
+namespace Sourceability\InstrumentationBundle\Bundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
