+++ src/Bundle/Resources/config/profiler_newrelic.yaml
