+++ src/Bundle/Resources/config/profiler_symfony.yaml
