--- src/Resources/config/listener_command.yaml
@@ -1,7 +1,0 @@
-services:
-    Sourceability\InstrumentationBundle\EventListener\CommandProfilerListener:
-        class: 'Sourceability\InstrumentationBundle\EventListener\CommandProfilerListener'
-        arguments:
-            - '@Sourceability\InstrumentationBundle\Profiler\ProfilerInterface'
-        tags:
-            - { name: 'kernel.event_subscriber' }
