--- src/Resources/config/listener_messenger.yaml
@@ -1,7 +1,0 @@
-services:
-    Sourceability\InstrumentationBundle\EventListener\MessengerProfilerListener:
-        class: 'Sourceability\InstrumentationBundle\EventListener\MessengerProfilerListener'
-        arguments:
-            - '@Sourceability\InstrumentationBundle\Profiler\ProfilerInterface'
-        tags:
-            - { name: 'kernel.event_subscriber' }
